+++ packages/core/shared/types.ts
@@ -112,7 +112,10 @@ export interface Message {
   tool_calls?: string | ToolCall[];
 
   /**
+
    * Additional message-specific information added on the server via StreamData 
+  * Additional message-specific information added on the server via StreamData
+
    */
   annotations?: JSONValue[] | undefined;
 }
@@ -312,8 +315,13 @@ export type UseCompletionOptions = {
    * })
    * ```
    */
+
+  body?: object
+}
+
   body?: object;
 };
+
 
 export type JSONValue =
   | null
@@ -321,6 +329,9 @@ export type JSONValue =
   | number
   | boolean
   | { [x: string]: JSONValue }
+
+  | Array<JSONValue>
+
   | Array<JSONValue>;
 
 export type AssistantMessage = {
@@ -346,3 +357,4 @@ export type DataMessage = {
   role: 'data';
   data: JSONValue; // application-specific data
 };
+