+++ packages/core/streams/huggingface-stream.ts
@@ -1,9 +1,16 @@
 import {
+
+  type AIStreamCallbacks,
+  createCallbacksAndOptionsTransformer,
+  trimStartOfStreamHelper
+} from './ai-stream'
+=======
   type AIStreamCallbacksAndOptions,
   createCallbacksTransformer,
   trimStartOfStreamHelper,
 } from './ai-stream';
 import { createStreamDataTransformer } from './stream-data';
+
 
 function createParser(res: AsyncGenerator<any>) {
   const trimStartOfStream = trimStartOfStreamHelper();
@@ -40,9 +47,13 @@ export function HuggingFaceStream(
   res: AsyncGenerator<any>,
   callbacks?: AIStreamCallbacksAndOptions,
 ): ReadableStream {
+
+  return createParser(res).pipeThrough(createCallbacksAndOptionsTransformer(callbacks))
+=======
   return createParser(res)
     .pipeThrough(createCallbacksTransformer(callbacks))
     .pipeThrough(
       createStreamDataTransformer(callbacks?.experimental_streamData),
     );
+
 }